move Hy3-specific merge scripts and MTP utilities to tools/hy3
