+++ tools/hy3/merge_hy3_nvfp4_c8perhead.py
